requirements.txt
diffusers
matplotlib==3.10.8
numpy==2.4.2
pandas==2.3.3
Pillow==12.0.0
scikit-dimension==0.3.4
scikit-learn==1.8.0
seaborn==0.13.2
timm
torch==2.10.0
torchvision==0.25.0
transformers==5.2.0
